extract plotly and generic utils to own files
